requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
python-dotenv==1.0.1
snowflake-connector-python==3.10.1
